# docs/api/openapi-v1.yaml
openapi: 3.1.0
info:
  title: ATape HTTP API
  version: 1.0.0
  license:
    name: MIT
    identifier: MIT
  description: |
    First-party ATape Web and CLI contract. Domain behavior is owned by the
    Authentication, Team, ingestion, conversation, Raw, and Search Modules;
    this document describes only their HTTP representation.
servers:
  - url: https://api.atape.dev
    description: Official service API origin
...
paths:
...
  /api/v1/teams:
    post:
      operationId: createTeam
      tags: [Teams]
      summary: Create a Team with the current User as Owner
      x-atape-route: {class: web_only, requestBodyPolicy: json-control-plane, requestBodyLimit: 65536, cachePolicy: no-store, originPolicy: exact-web, freshAuthentication: false, requiresIdempotencyKey: true}
      security: [{webSession: []}]
      parameters:
        - {$ref: "#/components/parameters/Origin"}
        - {$ref: "#/components/parameters/CSRF"}
        - {$ref: "#/components/parameters/IdempotencyKey"}
      requestBody: {$ref: "#/components/requestBodies/TeamCreate"}
      responses:
        "201": {$ref: "#/components/responses/Team"}
        "4XX": {$ref: "#/components/responses/Problem"}
        default: {$ref: "#/components/responses/Problem"}
  /api/v1/teams/{teamSlug}:
    parameters:
      - {$ref: "#/components/parameters/TeamSlug"}
    get:
      operationId: getTeam
      tags: [Teams]
      summary: Read visible Team metadata
      x-atape-route: {class: any_principal, requestBodyPolicy: none, requestBodyLimit: 0, cachePolicy: no-store, originPolicy: none, freshAuthentication: false, requiresIdempotencyKey: false}
      security: [{webSession: []}, {cliCredential: []}]
      responses:
        "200": {$ref: "#/components/responses/Team"}
        "4XX": {$ref: "#/components/responses/Problem"}
        default: {$ref: "#/components/responses/Problem"}
    patch:
      operationId: updateTeam
      tags: [Teams]
      summary: Update a Team display profile as Owner
      x-atape-route: {class: web_only, requestBodyPolicy: json-control-plane, requestBodyLimit: 65536, cachePolicy: no-store, originPolicy: exact-web, freshAuthentication: false, requiresIdempotencyKey: false}
      security: [{webSession: []}]
      parameters:
        - {$ref: "#/components/parameters/Origin"}
        - {$ref: "#/components/parameters/CSRF"}
      requestBody: {$ref: "#/components/requestBodies/TeamUpdate"}
      responses:
        "200": {$ref: "#/components/responses/Team"}
        "4XX": {$ref: "#/components/responses/Problem"}
        default: {$ref: "#/components/responses/Problem"}
...
  /api/v1/teams/{teamSlug}/leave:
    parameters:
      - {$ref: "#/components/parameters/TeamSlug"}
    post:
      operationId: leaveTeam
      tags: [Teams]
      summary: Remove the current User's Team Membership
      x-atape-route: {class: web_only, requestBodyPolicy: json-control-plane, requestBodyLimit: 65536, cachePolicy: no-store, originPolicy: exact-web, freshAuthentication: false, requiresIdempotencyKey: false}
      security: [{webSession: []}]
      parameters:
        - {$ref: "#/components/parameters/Origin"}
        - {$ref: "#/components/parameters/CSRF"}
      requestBody: {$ref: "#/components/requestBodies/EmptyJSON"}
      responses:
        "204": {$ref: "#/components/responses/NoContent"}
        "4XX": {$ref: "#/components/responses/Problem"}
        default: {$ref: "#/components/responses/Problem"}
...
  /api/v1/team-memberships:
    post:
      operationId: joinTeam
      tags: [Teams]
      summary: Join or restore Membership using a Join Code
      x-atape-route: {class: web_only, requestBodyPolicy: json-control-plane, requestBodyLimit: 65536, cachePolicy: no-store, originPolicy: exact-web, freshAuthentication: false, requiresIdempotencyKey: false}
      security: [{webSession: []}]
      parameters:
        - {$ref: "#/components/parameters/Origin"}
        - {$ref: "#/components/parameters/CSRF"}
      requestBody: {$ref: "#/components/requestBodies/TeamJoin"}
      responses:
        "200": {$ref: "#/components/responses/Team"}
        "201": {$ref: "#/components/responses/Team"}
        "4XX": {$ref: "#/components/responses/Problem"}
        default: {$ref: "#/components/responses/Problem"}
  /api/v1/project-matches:
    post:
      operationId: matchProject
      tags: [Projects]
      summary: Match a normalized Git remote within one visible Team
      x-atape-route: {class: cli_only, requestBodyPolicy: json-control-plane, requestBodyLimit: 65536, cachePolicy: no-store, originPolicy: none, freshAuthentication: false, requiresIdempotencyKey: false}
      security: [{cliCredential: []}]
      requestBody: {$ref: "#/components/requestBodies/ProjectMatch"}
      responses:
        "200": {$ref: "#/components/responses/ProjectMatch"}
        "4XX": {$ref: "#/components/responses/Problem"}
        default: {$ref: "#/components/responses/Problem"}
  /api/v1/teams/{teamSlug}/projects:
    parameters:
      - {$ref: "#/components/parameters/TeamSlug"}
    post:
      operationId: createProject
      tags: [Projects]
      summary: Create a Git or Folder Project
      x-atape-route: {class: any_principal, requestBodyPolicy: json-control-plane, requestBodyLimit: 65536, cachePolicy: no-store, originPolicy: exact-web-when-cookie, freshAuthentication: false, requiresIdempotencyKey: true}
      security: [{webSession: []}, {cliCredential: []}]
      parameters:
        - {$ref: "#/components/parameters/OriginOptionalForCLI"}
        - {$ref: "#/components/parameters/CSRFOptionalForCLI"}
        - {$ref: "#/components/parameters/IdempotencyKey"}
      requestBody: {$ref: "#/components/requestBodies/ProjectCreate"}
      responses:
        "201": {$ref: "#/components/responses/Project"}
        "4XX": {$ref: "#/components/responses/Problem"}
        default: {$ref: "#/components/responses/Problem"}
  /api/v1/projects/{projectId}:
    parameters:
      - {$ref: "#/components/parameters/ProjectId"}
    get:
      operationId: getProject
      tags: [Projects]
      summary: Read visible Project metadata
      x-atape-route: {class: any_principal, requestBodyPolicy: none, requestBodyLimit: 0, cachePolicy: no-store, originPolicy: none, freshAuthentication: false, requiresIdempotencyKey: false}
      security: [{webSession: []}, {cliCredential: []}]
      responses:
        "200": {$ref: "#/components/responses/Project"}
        "4XX": {$ref: "#/components/responses/Problem"}
        default: {$ref: "#/components/responses/Problem"}
    patch:
      operationId: renameFolderProject
      tags: [Projects]
      summary: Rename a Folder Project as Owner
      x-atape-route: {class: web_only, requestBodyPolicy: json-control-plane, requestBodyLimit: 65536, cachePolicy: no-store, originPolicy: exact-web, freshAuthentication: false, requiresIdempotencyKey: false}
      security: [{webSession: []}]
      parameters:
        - {$ref: "#/components/parameters/Origin"}
        - {$ref: "#/components/parameters/CSRF"}
      requestBody: {$ref: "#/components/requestBodies/ProjectRename"}
      responses:
        "200": {$ref: "#/components/responses/Project"}
        "4XX": {$ref: "#/components/responses/Problem"}
        default: {$ref: "#/components/responses/Problem"}
    delete:
      operationId: deleteProject
      tags: [Projects]
      summary: Delete a Project and its captured data as fresh Owner
      x-atape-route: {class: web_only, requestBodyPolicy: none, requestBodyLimit: 0, cachePolicy: no-store, originPolicy: exact-web, freshAuthentication: true, requiresIdempotencyKey: false}
      security: [{webSession: []}]
      parameters:
        - {$ref: "#/components/parameters/Origin"}
        - {$ref: "#/components/parameters/CSRF"}
      responses:
        "204": {$ref: "#/components/responses/NoContent"}
        "4XX": {$ref: "#/components/responses/Problem"}
        default: {$ref: "#/components/responses/Problem"}
  /api/v1/projects/{projectId}/repository-relinks:
    parameters:
      - {$ref: "#/components/parameters/ProjectId"}
    post:
      operationId: relinkGitProject
      tags: [Projects]
      summary: Relink a Git Project remote as fresh Owner
      x-atape-route: {class: web_only, requestBodyPolicy: json-control-plane, requestBodyLimit: 65536, cachePolicy: no-store, originPolicy: exact-web, freshAuthentication: true, requiresIdempotencyKey: false}
      security: [{webSession: []}]
      parameters:
        - {$ref: "#/components/parameters/Origin"}
        - {$ref: "#/components/parameters/CSRF"}
      requestBody: {$ref: "#/components/requestBodies/ProjectRelink"}
      responses:
        "200": {$ref: "#/components/responses/Project"}
        "4XX": {$ref: "#/components/responses/Problem"}
        default: {$ref: "#/components/responses/Problem"}
  /api/v1/projects/{projectId}/archive:
    parameters:
      - {$ref: "#/components/parameters/ProjectId"}
    post:
      operationId: archiveProject
      tags: [Projects]
      summary: Idempotently archive a Project as Owner
      x-atape-route: {class: web_only, requestBodyPolicy: json-control-plane, requestBodyLimit: 65536, cachePolicy: no-store, originPolicy: exact-web, freshAuthentication: false, requiresIdempotencyKey: false}
      security: [{webSession: []}]
      parameters:
        - {$ref: "#/components/parameters/Origin"}
        - {$ref: "#/components/parameters/CSRF"}
      requestBody: {$ref: "#/components/requestBodies/EmptyJSON"}
      responses:
        "200": {$ref: "#/components/responses/Project"}
        "4XX": {$ref: "#/components/responses/Problem"}
        default: {$ref: "#/components/responses/Problem"}
...
components:
  securitySchemes:
    webSession:
      type: apiKey
      in: cookie
      name: __Secure-atape_session
      description: The production Cookie name is derived from topology and may use the __Host- prefix.
    cliCredential:
      type: http
      scheme: bearer
      bearerFormat: atc_v1 opaque credential
  parameters:
    Origin:
      name: Origin
      in: header
      required: true
      schema: {type: string, format: uri}
      description: Must exactly equal the configured canonical Web origin.
    OriginOptionalForCLI:
      name: Origin
      in: header
      required: false
      schema: {type: string, format: uri}
      description: Required and exact when the operation uses a Web Session; omitted for CLI Credential calls.
    CSRF:
      name: X-ATape-CSRF
      in: header
      required: true
      schema: {type: string, minLength: 1}
      description: Session-bound synchronizer token from the Session bootstrap response.
    CSRFOptionalForCLI:
      name: X-ATape-CSRF
      in: header
      required: false
      schema: {type: string, minLength: 1}
      description: Required when the operation uses a Web Session; omitted for CLI Credential calls.
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: {type: string, minLength: 22, maxLength: 128}
      description: A caller-generated 128-bit random operation key, retained for 24 hours.
    TeamSlug:
      name: teamSlug
      in: path
      required: true
      schema: {type: string, minLength: 1, maxLength: 63}
...
    ProjectId:
      name: projectId
      in: path
      required: true
      schema: {$ref: "#/components/schemas/OpaqueId"}
...
  requestBodies:
    EmptyJSON:
      required: true
      content:
        application/json:
          schema: {$ref: "#/components/schemas/EmptyObject"}
...
    TeamCreate:
      required: true
      content:
        application/json:
          schema:
            type: object
            additionalProperties: false
            required: [slug, displayName]
            properties:
              slug: {type: string, minLength: 1, maxLength: 63}
              displayName: {type: string, minLength: 1, maxLength: 200}
    TeamUpdate:
      required: true
      content:
        application/json:
          schema:
            type: object
            additionalProperties: false
            required: [displayName]
            properties:
              displayName: {type: string, minLength: 1, maxLength: 200}
...
    TeamJoin:
      required: true
      content:
        application/json:
          schema:
            type: object
            additionalProperties: false
            required: [joinCode]
            properties:
              joinCode: {type: string, minLength: 6, maxLength: 16}
    ProjectMatch:
      required: true
      content:
        application/json:
          schema:
            type: object
            additionalProperties: false
            required: [teamId, type, remote]
            properties:
              teamId: {$ref: "#/components/schemas/OpaqueId"}
              type: {const: git}
              remote: {type: string, minLength: 1}
    ProjectCreate:
      required: true
      content:
        application/json:
          schema:
            oneOf:
              - type: object
                additionalProperties: false
                required: [type, remote]
                properties:
                  type: {const: git}
                  remote: {type: string, minLength: 1}
              - type: object
                additionalProperties: false
                required: [type, name]
                properties:
                  type: {const: folder}
                  name: {type: string, minLength: 1, maxLength: 200}
    ProjectRename:
      required: true
      content:
        application/json:
          schema:
            type: object
            additionalProperties: false
            required: [name]
            properties:
              name: {type: string, minLength: 1, maxLength: 200}
    ProjectRelink:
      required: true
      content:
        application/json:
          schema:
            type: object
            additionalProperties: false
            required: [remote]
            properties:
              remote: {type: string, minLength: 1}
...
  responses:
    Problem:
      description: RFC 9457 Problem Details. The stable code determines client behavior.
      headers:
        X-Request-ID: {$ref: "#/components/headers/RequestId"}
        Cache-Control: {$ref: "#/components/headers/NoStore"}
        Retry-After: {$ref: "#/components/headers/RetryAfter"}
      content:
        application/problem+json:
          schema: {$ref: "#/components/schemas/Problem"}
    NoContent:
      description: The desired state has been reached.
      headers:
        X-Request-ID: {$ref: "#/components/headers/RequestId"}
        Cache-Control: {$ref: "#/components/headers/NoStore"}
...
    Team:
      description: Team metadata plus the caller's Membership role.
      headers:
        X-Request-ID: {$ref: "#/components/headers/RequestId"}
        Cache-Control: {$ref: "#/components/headers/NoStore"}
      content:
        application/json:
          schema: {$ref: "#/components/schemas/Team"}
...
    Project:
      description: Project metadata.
      headers:
        X-Request-ID: {$ref: "#/components/headers/RequestId"}
        Cache-Control: {$ref: "#/components/headers/NoStore"}
      content:
        application/json:
          schema: {$ref: "#/components/schemas/Project"}
    ProjectMatch:
      description: Exact same-Team Git match or none.
      headers:
        X-Request-ID: {$ref: "#/components/headers/RequestId"}
        Cache-Control: {$ref: "#/components/headers/NoStore"}
      content:
        application/json:
          schema: {$ref: "#/components/schemas/ProjectMatch"}
  headers:
    RequestId:
      required: true
      schema: {type: string, format: uuid}
    NoStore:
      required: true
      schema: {type: string, const: no-store}
...
    RetryAfter:
      required: false
      schema: {type: integer, minimum: 1}
      description: Present for retryable 409, 429, and 503 outcomes.
  schemas:
    EmptyObject:
      type: object
      additionalProperties: false
      maxProperties: 0
    OpaqueId:
      type: string
      minLength: 1
      maxLength: 200
      description: Opaque identifier without embedded authority.
    Timestamp:
      type: string
      format: date-time
...
    TeamMembershipSummary:
      type: object
      additionalProperties: false
      required: [role]
      properties:
        role: {type: string, enum: [owner, member]}
    Team:
      type: object
      additionalProperties: false
      required: [id, slug, displayName, membership, createdAt, updatedAt]
      properties:
        id: {$ref: "#/components/schemas/OpaqueId"}
        slug: {type: string}
        displayName: {type: string}
        membership: {$ref: "#/components/schemas/TeamMembershipSummary"}
        createdAt: {$ref: "#/components/schemas/Timestamp"}
        updatedAt: {$ref: "#/components/schemas/Timestamp"}
    Project:
      type: object
      additionalProperties: false
      required: [id, teamId, type, name, state, createdAt, updatedAt]
      properties:
        id: {$ref: "#/components/schemas/OpaqueId"}
        teamId: {$ref: "#/components/schemas/OpaqueId"}
        type: {type: string, enum: [git, folder]}
        name: {type: string}
        state: {type: string, enum: [active, archived]}
        repositoryIdentity: {type: string}
        capturedThrough: {$ref: "#/components/schemas/Timestamp"}
        createdAt: {$ref: "#/components/schemas/Timestamp"}
        updatedAt: {$ref: "#/components/schemas/Timestamp"}
...
    ProjectMatch:
      oneOf:
        - type: object
          additionalProperties: false
          required: [status]
          properties:
            status: {const: none}
        - type: object
          additionalProperties: false
          required: [status, project]
          properties:
            status: {const: exact}
            project: {$ref: "#/components/schemas/Project"}
...
    ProblemField:
      type: object
      additionalProperties: false
      required: [field, code]
      properties:
        field: {type: string}
        code: {type: string}
    Problem:
      type: object
      additionalProperties: true
      required: [type, title, status, code, detail, instance, requestId]
      properties:
        type: {type: string, format: uri, pattern: "^https://atape[.]dev/problems/v1/"}
        title: {type: string}
        status: {type: integer, minimum: 400, maximum: 599}
        code:
          type: string
          enum:
            - invalid_request
            - invalid_user_code
            - invalid_join_code
            - invalid_device_code
            - unsupported_protocol_version
            - ambiguous_credentials
            - unauthenticated
            - session_expired
            - session_revoked
            - fresh_authentication_required
            - origin_rejected
            - csrf_rejected
            - access_denied
            - membership_role_denied
            - credential_capability_denied
            - not_found
            - method_not_allowed
            - external_identity_conflict
            - last_owner_required
            - grant_already_decided
            - grant_consumed
            - idempotency_conflict
            - idempotency_in_progress
            - resource_state_conflict
            - login_expired
            - expired_token
            - request_too_large
            - unsupported_media_type
            - validation_failed
            - slow_down
            - too_many_code_attempts
            - internal_error
            - provider_unavailable
            - service_unavailable
        detail: {type: string}
        instance: {type: string, format: uri, pattern: "^urn:atape:request:"}
        requestId: {type: string, format: uuid}
        errors:
          type: array
          items: {$ref: "#/components/schemas/ProblemField"}